+++ RELEASE_NOTES.md
@@ -3,7 +3,7 @@
 This file contains one line summaries of commits that are worthy of mentioning in release notes.
 A new header is inserted each time a *tag* is created.
 
-- Changed the getWorldPosition() shading API to be less surprising.
+- Changed the user-facing getWorldPosition() shading API to return world-space coordinates.
 - Added support for vertex morphing.
 - Introduced custom attributes, accessible from the vertex shader.
 - Added Java / Kotlin bindings for KtxLoader.